Update to fmt 8.1.1 and fix enum printing
You aren't meant to print enum class values, there was a bug in older
versions of the fmt library that allowed us to. That has now been
fixed so we need to cast the value so we can print it.

+++ src/isotope.cpp
@@ -67,7 +67,7 @@ std::string Isotope::writeAsJSON(const bool human_readable) const
                      nubase.N,
                      nubase.symbol,
                      nubase.decay,
-                     nubase.exp,
+                     static_cast<uint8_t>(nubase.exp),
                      Converter::FloatToNdp(nubase.mass_excess, NDP),
                      Converter::FloatToNdp(nubase.dmass_excess, NDP),
                      Converter::FloatToNdp(ame.mass_excess, NDP),
